+++ lib/ruby_saml/idp_metadata_parser.rb
@@ -8,7 +8,6 @@ require "rexml/xpath"
 
 # Only supports SAML 2.0
 module RubySaml
-  include REXML
 
   # Auxiliary class to retrieve and parse the Identity Provider Metadata
   #
@@ -399,7 +398,7 @@ module RubySaml
 
           cert = OpenSSL::X509::Certificate.new(Base64.decode64(certificate))
 
-          fingerprint_alg = RubySaml::XML::BaseDocument.new.algorithm(fingerprint_algorithm).new
+          fingerprint_alg = RubySaml::XML::Crypto.hash_algorithm(fingerprint_algorithm).new
           fingerprint_alg.hexdigest(cert.to_der).upcase.scan(/../).join(":")
         end
       end